+++ src/vmp/kvlist.c
@@ -76,36 +76,42 @@ int kvlist_write(char* libpath, mt_map_t* db)
 	mt_vector_t* vals = VNEW(); // REL 1
 	mt_map_values(db, vals);
 
-	for (int vali = 0; vali < vals->length; vali++)
+	for (size_t vali = 0; vali < vals->length; vali++)
 	{
 	    mt_map_t*    entry = vals->data[vali];
 	    mt_vector_t* keys  = VNEW(); // REL 2
 
 	    mt_map_keys(entry, keys);
 
-	    for (int keyi = 0; keyi < keys->length; keyi++)
+	    for (size_t keyi = 0; keyi < keys->length; keyi++)
 	    {
 		char* key = keys->data[keyi];
 		char* val = MGET(entry, key);
 
-		if (fprintf(file, "%s\n", key) < 0) retv = -1;
-		if (fprintf(file, "%s\n", val) < 0) retv = -1;
+		if (fprintf(file, "%s\n", key) < 0)
+		    retv = -1;
+		if (fprintf(file, "%s\n", val) < 0)
+		    retv = -1;
 	    }
 
-	    if (fprintf(file, "-\n") < 0) retv = -1;
+	    if (fprintf(file, "-\n") < 0)
+		retv = -1;
 
 	    REL(keys); // REL 2
 
-	    if (retv < 0) break;
+	    if (retv < 0)
+		break;
 	}
 
-	if (fclose(file) == EOF) retv = -1; // CLOSE 0
+	if (fclose(file) == EOF)
+	    retv = -1; // CLOSE 0
 
 	REL(vals); // REL 1
 
 	if (retv == 0)
 	{
-	    if (rename(path, libpath) != 0) retv = -1;
+	    if (rename(path, libpath) != 0)
+		retv = -1;
 	}
 	else
 	    mt_log_error("ERROR kvlist_write cannot write file");
